get-friends-recently-watched
och lite fix i get-activities och en ny funktion i functions.
